new world: w_1_1_2 and w_1_3_2. transition between world is now possible

+++ super-mario-3-se102/Animation.cpp
@@ -42,7 +42,7 @@ void Animation::NoLoopUpdate(float delta) {
 
 	int i = int(currentDuration / frameDuration);
 	currentDuration -= frameDuration * i;
-	currentFrame = min(currentFrame + i, sequence.size() - 1);
+	currentFrame = min(currentFrame + i, (int)sequence.size() - 1);
 }
 
 const SpriteBox& Animation::GetCurrentSpriteBox()
